requirements.txt
dotenv
torch
transformers
redis
rq
pillow 
pypdfium2
fastapi[all]